#9: Completed implementation of all communication-related apparatus
With this commit everything regarding the inter-processor edges
is computed and kept up-to-date. Statistics were also modified to
cover the case of empty sets (as it is legal to have an empty set
of edge links).
We note that this communication-awareness does not yet include
taking it into account in the objective function. However, all
reports regarding communications are now correct.