#9: Implemented a communication-based cost function
With this first implementation of a communication-aware (and only
that) criterion, we now have the opposite extreme of the pure
load-balancing criteria (original Grapevine and modified 6' Grapevine).
In contrast the  2 latter, the new comm-aware criterion is "the other
extreme" in that it explicitly ignores workloads,  but instead, only
uses structure of the object-to-object communicators to make a
decision as to whether a transfer may or may not occur.

With this version we shall now be able to conduct performance
comparisons, between purely load-based approaches, versus
a hybrid approach where the main loop remains indeed a load-
balancing process (as task movement attempts are still
triggered by processor imbalance) but the decision to accept
or reject is only based on communication edges.

Future development will focus on changing the main optimization
loop to also make it communication-aware. Also, hybrid
load/communication concrete criteria will be created.