#52: complete overhaul of destination load usage and computation
The concept of "cached load" was eliminated. Instead, by default
the destination loads know to the senders are used, and updated
explicitly when an object is transferred.
As a result the dst load information is less outdated than when
cached, and yet does not require communication with destination
processors.
This option is on by default.
Load-based criteria were updated accordingly